Refactor interface to benchmarkSynchronousOperation.
This commit corrects a design flaw in the Perf interface, which caused
multiple problems.
1. There are too many variations of the benchmarkSynchronousOperation
   method to account for different combinations of arguments.
2. Adding arguments to benchmarkSynchronousOperation is not scalable,
   because doing so would either multiply the number of variations or
   require all callers to update their code.
3. Individual tests were manually converting parsed command line options
   into arguments for benchmarkSynchronousOperation, rather than doing
   that in a central place. This means that individual tests need to be
   updated to take advantage of new options.

This commit refactors benchmarkSynchronousOperation to use the
PerfArguments wrapper, and removes the extraneous variations of the
method.

It also simplifies the interface of Perf.runBenchmark, so that callers
can just pass their operation, the command line options object, and the
desired timestamp.

It also adds targetQPS and maxThreads to the PerfUtils interface, but
these are not implemented yet.