Handle descending ranges in PerfUtils.parseRanges.
Previously, if you wanted to do a threadRange test in which you used
fewer and fewer threads, you would have to specify the individual thread
counts.

./perf.sh NoOpThroughputShortDuration -threadRange 10,9,8,7,6,5,4,3,2,1

This is not symmetric with the other direction, where you could specify
a range.

./perf.sh NoOpThroughputShortDuration -threadRange 1-10

This commit allows the range to work in both directions, so that the
following invocations are valid:

./perf.sh NoOpThroughputShortDuration -threadRange 10-1

./perf.sh NoOpThroughputShortDuration -threadRange 10-1:2

./perf.sh NoOpThroughputShortDuration -threadRange 1-10,9-1