Add option enableGracefulTermination.
When specified, this option will install a shutdown hook so that
Control-C will cause Perf to compute and output statistics for the
experiment up to the current point in time before termination.

This option is not turned on globally or by default because it can
prevent Control-C from killing the application quickly, even when
invoked multiple times.