Add FailureCaptureTest into TestInfo.
This commit also modifies the test to support optional failure mapping,
which allows the test to be used for validating that failure mapping
interacts correctly with uncaught exceptions as well as verbosity=3.

The following invocations are particularly useful to ensuring that
uncaught exceptions are not mapped by the failure mapper, while
test-generated failure descriptions are mapped by the failure mapper.

./perf.sh  FailureCaptureTest
./perf.sh  FailureCaptureTest -useFailureMapper
./perf.sh  FailureCaptureTest -rethrowExceptions
./perf.sh  FailureCaptureTest -useFailureMapper -rethrowExceptions
./perf.sh  FailureCaptureTest -useFailureMapper -verbosity=3
./perf.sh  FailureCaptureTest -useFailureMapper -rethrowExceptions -verbosity=3