Support maxOperations when numThreads > 1.
Prior to this commit, Perf did not support maxOperations for throughpout
tests with numThreads > 1, because doing so required thread
synchronization on the hot path, which is an expensive operation when
one is trying to measure fast operations.

The performance penalty of doing this hasn't gone away, but this commit
introduces the functionality in a way that introduces no additional
overhead when maxOperations is not used. Note that it *does* introduce
the atomic increment overhead for the single-threaded latency test, but
I think that is reasonable because the increment is done outside the
measurement loop.

Why I believe this implementation is reasonable and adding it makes
sense.
 * In the pure latency case, you are not generating load anyways, so you
   do not care about the atomic increment overhead. The performance
   penalty is also smaller because you're not pinging it back and forth
   between cores.
 * In the maxDuration case with multiple threads, you do not pay for the atomic
   increment if you do not use maxOperations.
 * In the maxDuration + maxOperations case, you do pay for that
   synchronization, but a perf test would have to pay for synchronization
   regardless of whether the library offered it if they wanted that
   behavior.